01_chat_completion_chatbot/requirements.txt
streamlit
OpenAI
python-dotenv